firmware/pic_capture/pyproject.toml
[project]
name = "pic-capture"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "picamera2>=0.3.24",
    "python-prctl>=1.8.1",
    "rpi-lgpio>=0.6",
]
